Move & Rename SnapPreview, SnapView
